requirements.txt
asgiref==3.2.10
Django==3.0.8
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
pytz==2020.1
sqlparse==0.3.1
